pyproject.toml
[project]
name = "ml-zoomcamp-2025-capstone-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "ipykernel>=7.1.0",
    "matplotlib>=3.10.8",
    "numpy>=2.4.1",
    "pandas>=2.3.3",
    "pillow>=12.1.0",
    "python-multipart>=0.0.21",
    "requests>=2.32.5",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "uvicorn>=0.40.0",
]
